+++ frontend/src/metabase/containers/SaveQuestionModal/SaveQuestionTitle.tsx
@@ -1,0 +1,9 @@
+import { useSaveQuestionContext } from "metabase/containers/SaveQuestionModal/context";
+import { getTitle } from "metabase/containers/SaveQuestionModal/util";
+
+export const SaveQuestionTitle = () => {
+  const { question, showSaveType, multiStep } = useSaveQuestionContext();
+  const cardType = question.type();
+
+  return getTitle(cardType, showSaveType, multiStep);
+};
