+++ frontend/src/metabase/containers/SaveQuestionModal/index.ts
@@ -1,2 +1,4 @@
 export * from "./SaveQuestionModal";
 export { SAVE_QUESTION_SCHEMA } from "metabase/containers/SaveQuestionModal/schema";
+export { LLMSuggestionQuestionInfo } from "metabase/containers/SaveQuestionModal/LLMSuggestionQuestionInfo";
+export { SaveQuestionTitle } from "metabase/containers/SaveQuestionModal/SaveQuestionTitle";
